Can now run the ExportItems ews operation.
Called from the Account class only because I couldn't figure where was best to put it. Calling Account.export() with a list of ItemId() objects will return a list of tuples, the first element being the Id, the second element being a string containing the export output.

Background on the export output. It's a 64bit conversion of everything needed to recreate the item, full fidelity. It can be read straight back in with UploadItems, no changes needed.