Remove incorrect hard-coding of "GMT" for "GMT Standard Time"
The hard-coding had the effect of picking the pytz timezone GMT for
events in the MS timezone "GMT Standard Time", which is wrong - we
need a pytz timezone where there is UTC+0 in winter and UTC+1 in
summer. Removing this mapping means that a more correct tz will be
selected.

This is a partial fix for #519

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>