Store the territory associated with each timezone mapping
For each CLDR timezone, there is a territory; one entry corresponding
to each Microsoft timezone is the default mapping, which is territory
'001' in addition to its usual territory. So store the territory in
the CLDR_TO_MS_TIMEZONE_MAP (keeping the 001 territory by
preference).

Then, when generating the reverse map MS_TIMEZONE_TO_PYTZ_MAP, take
the mapping with territory 001.

This is a partial fix for #519

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>