Ensure BaseFolder.glob() is case-insensitive (#736)
Exchange folder names appear to be case-insensitive (although there
isn't documentation defining that with certainty), but the arguments to
BaseFolder.glob() may be interpreted in a case-sensitive way depending
on the operating system.

glob() uses fnmatch.fnmatch(), which passes arguments through
os.path.normcase(), which behaves differently on Windows compared to
other operating systems.

This forces a case-insensitive comparison to match Exchange.